+++ Sources/StreamChat/Query/Filter+predicate.swift
@@ -172,7 +172,7 @@ extension Filter {
                 return nil
             }
             return NSPredicate(
-                format: "%K CONTAINS[c] %@".prepend(("ANY "), ifCondition: isCollectionFilter),
+                format: "%K CONTAINS[cd] %@".prepend(("ANY "), ifCondition: isCollectionFilter),
                 argumentArray: [keyPathString, prefix]
             )
         case .contains where mappedValue is String: